perf(python): never import hypothesis in user code
