+++ py-polars/polars/testing/__init__.py
@@ -1,0 +1,11 @@
+from polars.testing.asserts import (
+    assert_frame_equal,
+    assert_frame_equal_local_categoricals,
+    assert_series_equal,
+)
+
+__all__ = [
+    "assert_series_equal",
+    "assert_frame_equal",
+    "assert_frame_equal_local_categoricals",
+]
