+++ py-polars/tests/parametric/test_dataframe.py
@@ -7,7 +7,8 @@ from hypothesis import example, given, settings
 from hypothesis.strategies import integers
 
 import polars as pl
-from polars.testing import assert_frame_equal, column, dataframes
+from polars.testing import assert_frame_equal
+from polars.testing._parametric import column, dataframes
 
 
 @given(df=dataframes())
